requirements.txt
Python 
Streamlit
TensorFlow
NumPy
Matplotlib
Pillow